Chore - move remaining API steps to apiHandler file
